+++ db/migrate/20180914064834_add_rates.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class AddRates < ActiveCurrency::Migration
-  def up
-    ActiveCurrency::AddRates.new.call
-  end
-end